Minor cleaning up. Factored the interrupt save/restore code for fragments out into two generated functions. This de-clutters the code cache somewhat.
Experimented with using GlobalContext in the test cases in in/direct entry code instead of explicitly passing it through. I failed to realize that the test cases use many manually constructed contexts instead of the global context. This led to some small amount of confusion when I was seeing a segfault. Anyway, this suggests some flakiness in my design when considering the generality of the GlobalContext function. In practice, I think that it should be used very sparingly.