Added a new gdb command, print-kernel-modules, that prints out all loaded kernel module info.
Cleaned up the kernel tool a bit, and added the ability to include and exclude certain syscalls from instrumentation. Still TODO: improve the docstring of --attach_to_syscalls.

Added a new exception table entry handler for MOVZX_GPRv_MEMb. Found this when trying to ping various addresses when Granary was attached.

Noticed that the recvmsg syscall during instrumentation of the kernel always returned an error for a ping 10.0.0.7. Not sure if this is expected or not but it's potentially worth investigating.