Fixed some bugs related to tool dependencies not being correctly loaded.
Fixed some issues related to wrapping, and made it possible for multiple wrappers to apply to the same function.

Still to do: right now wrapping introduces the unfortunate side effect of execution going native when transparent returns are enabled and the wrapper returns. I'm not sure how to resolve this just yet. I think what I will have to do is inject code that modifies the return address, or something to this effect.