WIP. Test cases are failing because of a linker issue. What's happening is that some of the .init_array entries related to granary::internal::SlabAllocator objects are being move out of .granary_init_array and back into .init_array by the compiler/linker when linking against the googletest-based test cases. This generally makes sense, but is a bit unexpected too. In theory, there is no 'right' solution to this problem, but I find it unexpected that the section changes without warning. A partial fix might involve not tearing down the heap in os/linux/user/memory.cc, and eliding the IsMemoryInUse check in base/new.cc for test builds.
Started work on a generic base class for memory introspection. Also started refactoring other tools to use this. This eliminates a bunch of the copypasta. At the moment, this stuff won't work because it uses things like [%0 + %1] in inline assembly effective addresses. I still have to add support for this in the parser.

Refactored arch::Operand::mem.reg_base -> base, .reg_index -> index, to be VirtualRegister objects. This required a bunch of overall changes here and there, and I think things are generally in working order, but more thorough testing is absolutely needed. This is the first step in addressing Issue #65. From here, I will go on to inlining SSAInstruction into NativeInstruction, and make it into a sort of view of virtual registers. I think this will be good.

I think some structures have become a bit bloated. For example, CodeFragment is over 1KB, which is sort of ridiculous.